refactor: Extract truncate_table.rs into focused modules
Split 925-line truncate_table.rs into four focused modules:

- truncate/core.rs: Core truncate logic (truncate operations, sequence resetting, cascade coordination)
- truncate/triggers.rs: DELETE trigger validation
- truncate/constraints.rs: Foreign key constraint validation
- truncate/mod.rs: Public API and executor coordination

Benefits:
- Each module under 250 lines, focused single responsibility
- Easier to test individual concerns
- Clear separation: core logic vs triggers vs constraints
- All 27 unit tests pass (5 pre-existing failures in cascade tests unrelated to refactoring)
- Backward compatible via re-export in truncate_table.rs
- No public API changes
